main-example/pyproject.toml
[tool.poetry]
name = "main-example"
version = "0.1.0"
description = ""
authors = ["Hunter Henrichsen <hunter.henrichsen@gmail.com>"]
readme = "README.md"
packages = [{include = "main_example"}]

[tool.poetry.dependencies]
python = "^3.11"
flask = "^2.2.3"
dependency-injector = "^4.41.0"
pyjwt = "^2.6.0"
flask-restful = "^0.3.9"
python-dotenv = "^1.0.0"
argon2-cffi = "^21.3.0"
psycopg2 = "^2.9.5"
flask-jwt-extended = "^4.4.4"


[tool.poetry.group.dev.dependencies]
black = "^23.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
